+++ CHANGELOG.md
@@ -1,6 +1,18 @@
 Changelog
 ==========
 
+Version 6.17.3 *(2020-11-10)*
+----------------------------
+
+ * Properly update the folder thumbnail text colors
+
+Version 6.17.2 *(2020-11-09)*
+----------------------------
+
+ * Adding some folder thumbnail customization at the app settings
+ * Allow setting a default folder to be opened at app launch (by roland-kister)
+ * Updated the photo editor to fix some glitches
+
 Version 6.17.1 *(2020-11-03)*
 ----------------------------